- added maxlife to LifeBar - added maxMana to ManaBar - spell and hero properties changed

+++ src/main/java/com/khryniewicki/projectX/game/control_settings/mouse_settings/MouseSettings.java
@@ -78,13 +78,13 @@ public class MouseSettings {
     }
 
     private void consumeSpellMana() {
-        Integer heroMana = hero.getMana();
+        Float heroMana = hero.getMana();
         hero.setMana(heroMana - spell.getManaConsumed());
 
         ManaBar manaBar = hero.getManaBar();
         manaBar.updateManaBar();
-
         stackEvent.addHeroDto();
+
     }
