+++ src/main/java/com/khryniewicki/projectX/services/receive_services/ReceiveServiceSingleton.java
@@ -22,7 +22,7 @@ public class ReceiveServiceSingleton {
         return heroReceiveService.getMockLife();
     }
 
-    public Integer get_hero_mock_mana() {
+    public Float get_hero_mock_mana() {
         return heroReceiveService.getMockMana();
     }
